move requestmethod
single class packages don't make any sense.
